requirements.txt
streamlit
ultralytics
opencv-python
numpy
Pillow
kagglehub